+++ src/main/java/dev/orne/beans/converters/EnumConverter.java
@@ -25,47 +25,48 @@ package dev.orne.beans.converters;
 import javax.validation.constraints.NotNull;
 
 import org.apache.commons.beanutils.converters.AbstractConverter;
+import org.apache.commons.lang3.Validate;
 
 /**
  * Implementation of {@code Converter} that converts {@code Enum} instances
  * to and from {@code String} using value name as {@code String}
  * representation.
  * 
- * @param <E> The type of enumeration this instance converters
  * @author <a href="mailto:wamphiry@orne.dev">(w) Iker Hernaez</a>
- * @version 1.0, 2020-05
+ * @version 2.0, 2021-09
  * @since 0.1
  */
-public class EnumConverter<E extends Enum<E>>
+public class EnumConverter
 extends AbstractConverter {
 
-    /** The type of enumeration this instance converters. */
-    private final Class<E> enumType;
+    /**
+     * Shared instance that throws a {@code ConversionException} if an
+     * error occurs.
+     */
+    public static final EnumConverter GENERIC = new EnumConverter();
+    /**
+     * Shared instance that returns {@code null} if an error occurs.
+     */
+    public static final EnumConverter GENERIC_DEFAULT = new EnumConverter(null);
 
     /**
-     * Creates a new instance that throws a {@code ConversionException} if an
+     * Creates a new generic instance that throws a {@code ConversionException} if an
      * error occurs.
-     * 
-     * @param enumType The type of enumeration this instance converters
      */
-    public EnumConverter(
-            final @NotNull Class<E> enumType) {
+    private EnumConverter() {
         super();
-        this.enumType = enumType;
     }
 
     /**
-     * Creates a new instance that returns a default value if an error occurs.
+     * Creates a new generic instance that returns a default value if an error
+     * occurs.
      * 
-     * @param enumType The type of enumeration this instance converters
      * @param defaultValue The default value to be returned if the value to be
      * converted is missing or an error occurs converting the value
      */
-    public EnumConverter(
-            final @NotNull Class<E> enumType,
-            final E defaultValue) {
+    private EnumConverter(
+            final Enum<?> defaultValue) {
         super(defaultValue);
-        this.enumType = enumType;
     }
 
     /**
@@ -73,7 +74,7 @@ extends AbstractConverter {
      */
     @Override
     protected @NotNull Class<?> getDefaultType() {
-        return this.enumType;
+        return Enum.class;
     }
 
     /**
@@ -84,10 +85,44 @@ extends AbstractConverter {
             final @NotNull Class<T> type,
             final Object value)
     throws Throwable {
-        if (type.isAssignableFrom(this.enumType)) {
-            return type.cast(Enum.valueOf(this.enumType, value.toString()));
+        if (type.isEnum()) {
+            if (value == null) {
+                return null;
+            } else {
+                @SuppressWarnings("unchecked")
+                final Class<? extends Enum<?>> eType = (Class<? extends Enum<?>>) type;
+                return type.cast(enumFromName(eType, value.toString()));
+            }
         }
         throw conversionException(type, value);
+    }
+
+    /**
+     * Returns the enumeration constant of the specified type that matches the
+     * specified enumeration name.
+     * 
+     * @param <T> The enumeration type
+     * @param type The enumeration type
+     * @param name The name of the constant
+     * @return The enumeration constant for the name
+     * @throws ConversionException If the type is not an enumeration or no
+     * constant matches the specified name
+     */
+    protected <T> T enumFromName(
+            final @NotNull Class<T> type,
+            final @NotNull String name) {
+        Validate.notNull(type);
+        Validate.notNull(name);
+        final T[] constants = type.getEnumConstants();
+        if (constants == null) {
+            throw conversionException(type, name);
+        }
+        for (final T constant : constants) {
+            if (((Enum<?>) constant).name().equals(name)) {
+                return constant;
+            }
+        }
+        throw conversionException(type, name);
     }
 
     /**