+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 01. :boom: Changed `dev.orne.beans.UnrecognizedIdentityTokenException` hierarchy.
     Extends `IllegalArgumentException` now.
+01. :gift: Improved Apache `beanutils` enumerations converter
+    01. :boom: Removed generic nature of `EnumConverter` and exposed as singleton instance
+    01. :gift: Added class `dev.orne.beans.converters.EnumTypedConverter`
+        01. :gift: Added to method `OrneBeansConverters.registerBeansConversors(ConvertUtilsBean, ...)`
+    01. :gift: Added class `dev.orne.beans.converters.EnumConvertUtilsBean`
+        Extends `org.apache.commons.beanutils.ConvertUtilsBean` but lookups for
+        converter for `Enum` if no converter is found for concrete enumeration
+        type.
+    01. :gift: Added class `dev.orne.beans.converters.EnumConvertUtilsBean2`
+        Extends `org.apache.commons.beanutils.ConvertUtilsBean2` but lookups for
+        converter for `Enum` if no converter is found for concrete enumeration
+        type.
 
 # :package: 0.3.0